feat: bm 서브 카테고리 개수 조정 - fifth commit
- goalInvestment Long으로 타입 변경

+++ src/main/java/com/pitchain/dto/req/CreateBmReq.java
@@ -18,7 +18,7 @@ public record CreateBmReq(
         String description,
         String address,
         Long valuationCap,
-        Integer goalInvestment,
+        Long goalInvestment,
         Integer maxIssuedShare,
         LocalDate deadline,
         String longPitchUrl
