+++ src/test/java/com/pitchain/service/InvestmentServiceTest.java
@@ -102,7 +102,7 @@ class InvestmentServiceTest {
 
         long raisedAmount = investmentStatus.raisedAmount();
         assertThat(raisedAmount).isEqualTo(amountA + amountB);
-        Integer goalInvestment = investmentStatus.goalInvestment();
+        Long goalInvestment = investmentStatus.goalInvestment();
         assertThat(goalInvestment).isEqualTo(bm.getGoalInvestment());
         assertThat(investmentStatus.achievementRate()).isEqualTo(Math.min((int) ((raisedAmount / (double) goalInvestment) * 100), 100));
     }
@@ -114,6 +114,6 @@ class InvestmentServiceTest {
     private Bm saveBm(Member member) {
         return bmRepository.save(new Bm(member, "bmName", MainCategory.FOOD, "bmCompany", "logoImg",
                 "bmIntro", "bmDescription", "bmDescriptionImg", "companyAddress", 100000L,
-                1000, 1000, LocalDate.now(), "longPitchUrl"));
+                1000L, 1000, LocalDate.now(), "longPitchUrl"));
     }
 }
